// packages/shared/comms/v3/proto/broker_pb.d.ts
  export type AsObject = {
    type: MessageTypeMap[keyof MessageTypeMap],
  }
// ...
export class MessageHeader extends jspb.Message {
  getType(): MessageTypeMap[keyof MessageTypeMap];
  setType(value: MessageTypeMap[keyof MessageTypeMap]): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): MessageHeader.AsObject;
  static toObject(includeInstance: boolean, msg: MessageHeader): MessageHeader.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: MessageHeader, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): MessageHeader;
  static deserializeBinaryFromReader(message: MessageHeader, reader: jspb.BinaryReader): MessageHeader;
}

export namespace MessageHeader {
  export type AsObject = {
    type: MessageTypeMap[keyof MessageTypeMap],
  }
}

export class PingMessage extends jspb.Message {
  getType(): MessageTypeMap[keyof MessageTypeMap];
  setType(value: MessageTypeMap[keyof MessageTypeMap]): void;

  getTime(): number;
  setTime(value: number): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): PingMessage.AsObject;
  static toObject(includeInstance: boolean, msg: PingMessage): PingMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: PingMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): PingMessage;
  static deserializeBinaryFromReader(message: PingMessage, reader: jspb.BinaryReader): PingMessage;
}

export namespace PingMessage {
  export type AsObject = {
    type: MessageTypeMap[keyof MessageTypeMap],
    time: number,
  }
}

export class SubscriptionMessage extends jspb.Message {
  getType(): MessageTypeMap[keyof MessageTypeMap];
  setType(value: MessageTypeMap[keyof MessageTypeMap]): void;

  getFormat(): FormatMap[keyof FormatMap];
  setFormat(value: FormatMap[keyof FormatMap]): void;

  getTopics(): Uint8Array | string;
  getTopics_asU8(): Uint8Array;
  getTopics_asB64(): string;
  setTopics(value: Uint8Array | string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): SubscriptionMessage.AsObject;
  static toObject(includeInstance: boolean, msg: SubscriptionMessage): SubscriptionMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: SubscriptionMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): SubscriptionMessage;
  static deserializeBinaryFromReader(message: SubscriptionMessage, reader: jspb.BinaryReader): SubscriptionMessage;
}

export namespace SubscriptionMessage {
  export type AsObject = {
    type: MessageTypeMap[keyof MessageTypeMap],
    format: FormatMap[keyof FormatMap],
    topics: Uint8Array | string,
  }
}
// ...
export class TopicMessage extends jspb.Message {
  getType(): MessageTypeMap[keyof MessageTypeMap];
  setType(value: MessageTypeMap[keyof MessageTypeMap]): void;

  getFromAlias(): number;
  setFromAlias(value: number): void;

  getTopic(): string;
  setTopic(value: string): void;

  getBody(): Uint8Array | string;
  getBody_asU8(): Uint8Array;
  getBody_asB64(): string;
  setBody(value: Uint8Array | string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): TopicMessage.AsObject;
  static toObject(includeInstance: boolean, msg: TopicMessage): TopicMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: TopicMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): TopicMessage;
  static deserializeBinaryFromReader(message: TopicMessage, reader: jspb.BinaryReader): TopicMessage;
}

export namespace TopicMessage {
  export type AsObject = {
    type: MessageTypeMap[keyof MessageTypeMap],
    fromAlias: number,
    topic: string,
    body: Uint8Array | string,
  }
}

export class TopicFWMessage extends jspb.Message {
  getType(): MessageTypeMap[keyof MessageTypeMap];
  setType(value: MessageTypeMap[keyof MessageTypeMap]): void;

  getFromAlias(): number;
  setFromAlias(value: number): void;

  getBody(): Uint8Array | string;
  getBody_asU8(): Uint8Array;
  getBody_asB64(): string;
  setBody(value: Uint8Array | string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): TopicFWMessage.AsObject;
  static toObject(includeInstance: boolean, msg: TopicFWMessage): TopicFWMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: TopicFWMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): TopicFWMessage;
  static deserializeBinaryFromReader(message: TopicFWMessage, reader: jspb.BinaryReader): TopicFWMessage;
}

export namespace TopicFWMessage {
  export type AsObject = {
    type: MessageTypeMap[keyof MessageTypeMap],
    fromAlias: number,
    body: Uint8Array | string,
  }
}

export class TopicIdentityMessage extends jspb.Message {
  getType(): MessageTypeMap[keyof MessageTypeMap];
  setType(value: MessageTypeMap[keyof MessageTypeMap]): void;

  getFromAlias(): number;
  setFromAlias(value: number): void;

  getTopic(): string;
  setTopic(value: string): void;

  getIdentity(): Uint8Array | string;
  getIdentity_asU8(): Uint8Array;
  getIdentity_asB64(): string;
  setIdentity(value: Uint8Array | string): void;

  getRole(): RoleMap[keyof RoleMap];
  setRole(value: RoleMap[keyof RoleMap]): void;

  getBody(): Uint8Array | string;
  getBody_asU8(): Uint8Array;
  getBody_asB64(): string;
  setBody(value: Uint8Array | string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): TopicIdentityMessage.AsObject;
  static toObject(includeInstance: boolean, msg: TopicIdentityMessage): TopicIdentityMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: TopicIdentityMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): TopicIdentityMessage;
  static deserializeBinaryFromReader(message: TopicIdentityMessage, reader: jspb.BinaryReader): TopicIdentityMessage;
}

export namespace TopicIdentityMessage {
  export type AsObject = {
    type: MessageTypeMap[keyof MessageTypeMap],
    fromAlias: number,
    topic: string,
    identity: Uint8Array | string,
    role: RoleMap[keyof RoleMap],
    body: Uint8Array | string,
  }
}

export class TopicIdentityFWMessage extends jspb.Message {
  getType(): MessageTypeMap[keyof MessageTypeMap];
  setType(value: MessageTypeMap[keyof MessageTypeMap]): void;

  getFromAlias(): number;
  setFromAlias(value: number): void;

  getIdentity(): Uint8Array | string;
  getIdentity_asU8(): Uint8Array;
  getIdentity_asB64(): string;
  setIdentity(value: Uint8Array | string): void;

  getRole(): RoleMap[keyof RoleMap];
  setRole(value: RoleMap[keyof RoleMap]): void;

  getBody(): Uint8Array | string;
  getBody_asU8(): Uint8Array;
  getBody_asB64(): string;
  setBody(value: Uint8Array | string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): TopicIdentityFWMessage.AsObject;
  static toObject(includeInstance: boolean, msg: TopicIdentityFWMessage): TopicIdentityFWMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: TopicIdentityFWMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): TopicIdentityFWMessage;
  static deserializeBinaryFromReader(message: TopicIdentityFWMessage, reader: jspb.BinaryReader): TopicIdentityFWMessage;
}

export namespace TopicIdentityFWMessage {
  export type AsObject = {
    type: MessageTypeMap[keyof MessageTypeMap],
    fromAlias: number,
    identity: Uint8Array | string,
    role: RoleMap[keyof RoleMap],
    body: Uint8Array | string,
  }
}

export interface MessageTypeMap {
  UNKNOWN_MESSAGE_TYPE: 0;
  WELCOME: 1;
  CONNECT: 2;
  PING: 6;
  SUBSCRIPTION: 7;
  AUTH: 8;
  TOPIC: 9;
  TOPIC_FW: 10;
  TOPIC_IDENTITY: 11;
  TOPIC_IDENTITY_FW: 12;
}

export const MessageType: MessageTypeMap;

export interface RoleMap {
  UNKNOWN_ROLE: 0;
  CLIENT: 1;
  COMMUNICATION_SERVER: 2;
}

export const Role: RoleMap;

export interface FormatMap {
  UNKNOWN_FORMAT: 0;
  PLAIN: 1;
  GZIP: 2;
}

export const Format: FormatMap;